Added files
These two files are the part of the assignment that I wrote (mostly, some templates were used).

This does not work ~very~ well on it's own right now, but I plan to make this a standalone program. It is close, I just need maybe an hour to sort it through and get proper output based on input.

Other issues:
Issue with firstCase(size_t) algorithm - can you find it? (Ans: the first of the linked list may not be the first block in memory). 
Issue with findEntry(..........) algorithm - it is not an effective way to find the entry. I can do better with Hashes in Java, but is it worth implementing that in C?
Issue with myFree(.............) algorithm - will the code break if the block encapsulates the entire memory?